Moved file verification to repository layer
