+++ DotNetCore_UploadFile_Demo/Data/IImportDataResource.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Http;
+using System.Threading.Tasks;
+
+namespace DotNetCore_UploadFile_Demo.Data
+{
+    public interface IImportDataResource
+    {
+        Task<string> ValidateImportRequestData(HttpRequest request);
+    }
+}
